添加公共工具模块，重构窗口处理和错误显示逻辑
