+++ sub_adjust.lua
