+++ sub_adjust.py
@@ -1,11 +1,11 @@
 import os
 import re
-import platform
 import tkinter as tk
 from tkinter import (messagebox, scrolledtext)
 import webbrowser
-import winsound
 from docopt import docopt
+
+from common_utils import center_window, custom_messagebox, display_errors
 
 __version__ = 'sub-adjust v1.2.0'
 
@@ -158,50 +158,6 @@ def process_ass_ssa_file(filepath, adjusted_shift_value, layers):
         print(f"Error processing file {filepath}: {str(e)}")
         return False, str(e)
 
-
-def center_window(window):
-    window.update_idletasks()
-    screen_width = window.winfo_screenwidth()
-    screen_height = window.winfo_screenheight()
-    size = tuple(int(_) for _ in window.geometry().split('+')[0].split('x'))
-    x = screen_width // 2 - size[0] // 2
-    y = screen_height // 2 - size[1] // 2
-    window.geometry(f"{size[0]}x{size[1]}+{x}+{y}")
-
-def play_system_sound():
-    system = platform.system()
-    if system == "Windows":
-        winsound.MessageBeep(winsound.MB_OK)
-    elif system == "Linux":
-        pass
-    elif system == "Darwin":
-        pass
-
-def custom_messagebox(root, message):
-    play_system_sound()
-    custom_box = tk.Toplevel(root)
-    custom_box.title("处理结果")
-    tk.Label(custom_box, text=message, wraplength=300, justify=tk.LEFT).pack(padx=20, pady=20)
-    button_frame = tk.Frame(custom_box)
-    button_frame.pack(padx=10, pady=10)
-    tk.Button(button_frame, text="确定", command=custom_box.destroy).pack(side=tk.LEFT, padx=5)
-    tk.Button(button_frame, text="退出", command=root.destroy).pack(side=tk.LEFT, padx=5)
-    center_window(custom_box)
-    custom_box.transient(root)
-    custom_box.grab_set()
-    root.wait_window(custom_box)
-
-def display_errors(root, errors):
-    play_system_sound()
-    error_window = tk.Toplevel(root)
-    error_window.title("错误信息")
-    error_text = scrolledtext.ScrolledText(error_window, wrap=tk.CHAR, state='disabled')
-    error_text.pack(expand=True, fill='both')
-    error_text.config(state='normal')
-    error_text.insert(tk.END, errors)
-    error_text.config(state='disabled')
-    center_window(error_window)
-
 def shift_times_in_directory(directory, shift_value, shift_direction, layer_numbers):
     if shift_direction is None:
         shift_direction = "delay"
@@ -343,7 +299,7 @@ def main():
         print(__version__)
         return  # 退出程序，避免启动GUI
 
-    # 如果有offset参数，则执行命令行模式逻辑
+    # 如果有 -- offset 参数，则执行命令行模式逻辑
     elif args["--offset"]:
         print("进入命令行模式...")
         shift_value = float(args["--offset"]) * 1000
@@ -362,7 +318,7 @@ def main():
         else:
             shift_times_in_directory(os.getcwd(), shift_value, None, layer_numbers)
 
-    # 否则启动GUI（没有提供 --offset 或 --files 时）
+    # 否则启动GUI（没有提供 --offset 时）
     else:
         print("正在启动GUI...")
         start_ui()
